+++ src/keybinds/config.rs
@@ -118,7 +118,10 @@ pub struct LogTabKeybindsConfig {
     pub goto_child: Option<Keybind>,
     pub load_more: Option<Keybind>,
 
+    pub use_marks: Option<Keybind>,
+
     pub duplicate: Option<Keybind>,
+    pub parallelize: Option<Keybind>,
     pub create_new: Option<Keybind>,
     pub create_new_describe: Option<Keybind>,
     pub squash: Option<Keybind>,