+++ src/ui/op_log_tab.rs
@@ -216,7 +216,7 @@ impl Tab for OpLogTab<'_> {
         Ok(())
     }
 
-    fn open_context_menu(&self) -> Result<Option<AppAction>> {
+    fn open_context_menu(&mut self) -> Result<Option<AppAction>> {
         Ok(self.context_menu(self.op_panel.selected_position()))
     }