+++ src/ui/settings_tab.rs
@@ -343,7 +343,7 @@ impl Tab for SettingsTab {
         Ok(())
     }
 
-    fn open_context_menu(&self) -> Result<Option<AppAction>> {
+    fn open_context_menu(&mut self) -> Result<Option<AppAction>> {
         Ok(self.context_menu(
             self.settings_pane
                 .item_anchor(self.settings.selected_row(), 1),